Program Completed
*Renamed all files to fit naming conventions

*Changed Textformatter.java to work with key entries and accept
numbers only and insert them right to left up to a max value instead or
regex or the actual TextFormatter class. All methods that were dependent
on this class have been updated accordingly to leverage the class re-write.

*getMeals() had a major issue I found while traversing the array the nav
buttons where perpetually increasing the sizes of the arraylist's since they were using .add instead of .set the method has been updated to properly handle the 3 meal arrays and save/display them while using the the next day previous
day buttons and confirm button.

*Fixed at least one null pointer exception when navigating back to
getDateTime() and then attempting to move forward again which was not
prompting the dateTimeValid method to return false it will not return
properly and not throw the exception. I think I caused the other null pointer
exception while editing the TextFormatter class in the getMeals() which was
not an issue in previous commits. Either way now its fixed.